requirements.txt
requests
python-dotenv
sentence-transformers
faiss-cpu
fastapi
uvicorn
psutil
